requirement.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
constantly==15.1.0
cryptography==39.0.2
daphne==3.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
Pillow==9.4.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.3
pyOpenSSL==23.0.0
pytz==2022.7.1
redis==4.5.3
rest-framework-simplejwt==0.0.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.2.4
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
zope.interface==6.0
